fix : OAuth2 세션 저장 -> Jwt로 저장 (access Token, refresh Token)

+++ src/main/java/com/example/to_do_list/service/TodoService.java
@@ -7,6 +7,7 @@ import com.example.to_do_list.dto.todo.TodoResponsesDto;
 import com.example.to_do_list.dto.todo.TodoSaveDto;
 import com.example.to_do_list.dto.todo.TodoUpdateDto;
 import com.example.to_do_list.repository.TodoRepository;
+import com.example.to_do_list.repository.UsersRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
@@ -14,6 +15,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -21,17 +23,21 @@ import java.util.NoSuchElementException;
 @Service
 public class TodoService {
     private final TodoRepository todoRepository;
+    private final UsersRepository usersRepository;
 
-    public Long save(TodoSaveDto todoSaveDto) {
+    public Long save(TodoSaveDto todoSaveDto, Long usersId) {
         Todo todo = todoSaveDto.toEntity();
-//        todo.setUsers(users);
-//        users.addTodoList(todo);
+
+        Users users = findUsersById(usersId);
+        todo.setUsers(users);
+        users.addTodoList(todo);
         Todo saveTodo = todoRepository.save(todo);
 
         return saveTodo.getId();
     }
 
-    public Long update(Long id, TodoUpdateDto todoUpdateDto) {
+    public Long update(Long id, TodoUpdateDto todoUpdateDto, Long usersId) {
+        Users users = findUsersById(usersId);
         Todo todo = todoRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("존재하지 않는 할일입니다."));
 
@@ -49,9 +55,16 @@ public class TodoService {
                 .endDate(todo.getEndDate())
                 .build();
     }
-    public boolean changeStatus(Long id) {
+    public boolean changeStatus(Long id, Long usersId) {
+        Users users = findUsersById(usersId);
+        List<Todo> list = users.getTodoList();
+
         Todo todo = todoRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("todo not found"));
+        if(!list.contains(todo)) {
+            throw new IllegalArgumentException("자신의 todo만 변경 가능합니다.");
+        }
+
         return todo.updateStatus();
     }
     public Slice<TodoResponsesDto> findByDate(int page, int size, LocalDate date) {
@@ -60,12 +73,34 @@ public class TodoService {
         return todos;
     }
 
-    public void deleteTodo(Long id) {
+    public void deleteTodo(Long id, Long usersId) {
+        Users users = findUsersById(usersId);
+        List<Todo> list = users.getTodoList();
+        Todo todo = todoRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 todo"));
+        if(!list.contains(todo)) {
+            throw new IllegalArgumentException("본인의 todo만 삭제 가능");
+        }
         todoRepository.deleteById(id);
     }
 
-    public void deleteTodos(List<Long> ids) {
+    public void deleteTodos(List<Long> ids, Long usersId) {
+        Users users = findUsersById(usersId);
+        List<Long> list = new ArrayList<>();
+        List<Todo> todoList = users.getTodoList();
+        for (Todo todo : todoList) {
+            list.add(todo.getId());
+        }
+        for (Long id : ids) {
+            if(!list.contains(id)) {
+                throw new IllegalArgumentException("존재하지 않는 todoList");
+            }
+        }
         todoRepository.deleteAllById(ids);
     }
 
+    private Users findUsersById(Long usersId) {
+        return usersRepository.findById(usersId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 유저"));
+    }
 }
